commit add funcionalidades

+++ src/main/java/com/elo7/space_probe/ui/probes/ProbeDTO.java
@@ -1,3 +1,9 @@
 package com.elo7.space_probe.ui.probes;
 
-public record ProbeDTO(long id, String name, Integer x, Integer y, Integer planetId) { }
+public record ProbeDTO(
+        Integer id,
+        String name,
+        Integer x,
+        Integer y,
+        Integer planetId
+) { }
